Use EditorConfigDefaults to enable experimental rulesets
